remove deprecated data sources
