--- docs/data-sources/string.md
@@ -1,29 +1,0 @@
----
-# generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "oci_string Data Source - terraform-provider-oci"
-subcategory: ""
-description: |-
-  Data source for parsing a pinned oci string into its constituent parts. A pinned oci reference is one that includes a digest, and is in the format: '${registry}/${repo}@${digest}'. For example: 'cgr.dev/my-project/my-image@sha256:...'.
----
-
-# oci_string (Data Source)
-
-Data source for parsing a pinned oci string into its constituent parts. A pinned oci reference is one that includes a digest, and is in the format: '${registry}/${repo}@${digest}'. For example: 'cgr.dev/my-project/my-image@sha256:...'.
-
-
-
-<!-- schema generated by tfplugindocs -->
-## Schema
-
-### Required
-
-- `input` (String) The oci reference string to parse. This supports any valid oci reference string, including those with a tag, digest, or both. For example: 'cgr.dev/my-project/my-image:latest' or 'cgr.dev/my-project/my-image@sha256:...'. Note that when tags are provided, they will be replaced in favor of the digest.
-
-### Read-Only
-
-- `digest` (String) The digest of the oci reference. For example: `sha256:...`
-- `id` (String) The fully qualified oci reference string, in the format: '${registry}/${repo}@${digest}'. For example: `cgr.dev/my-project/my-image@sha256:...`
-- `pseudo_tag` (String) A pseudo tag pinned to a digest that can be used in place of a real tag. This is useful for cases where a tag is not provided, but is required for compatibility reasons. For example: `unused@sha256:...`. The tag always has the value `unused`, and the digest is the same as the input digest.
-- `registry` (String) The registry of the oci reference. For example: `cgr.dev`
-- `registry_repo` (String) Helper attribute equivalent to '${registry}/${repo}'. For example: `cgr.dev/my-project/my-image`
-- `repo` (String) The repository of the oci reference. For example: `my-project/my-image`
