--- examples/data-sources/oci_ref/data-source.tf
@@ -1,3 +1,0 @@
-data "oci_ref" "example" {
-  ref = "registry.example.com/repo/image:tag"
-}
